Move basiclayout to src root
