+++ lib/src/main/java/com/zetcode/basiclayout/CardLayoutEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.basiclayout;
 
 import java.awt.BorderLayout;
 import java.awt.CardLayout;
