+++ lib/src/main/java/com/zetcode/basiclayout/FlowLayoutEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.basiclayout;
 
 import javax.swing.JButton;
 import javax.swing.JFrame;
