+++ lib/src/main/java/com/zetcode/basiclayout/coffee2.png
