+++ lib/src/main/java/com/zetcode/basiclayout/computer.png
