+++ lib/src/main/java/com/zetcode/basiclayout/exit.png
