+++ lib/src/main/java/com/zetcode/basiclayout/next.png
