+++ lib/src/main/java/com/zetcode/basiclayout/previous.png
