+++ lib/src/main/java/com/zetcode/basiclayout/printer.png
